chore: Rename package `config` to `usrcfg` for better readability
